+++ modules/Date.js
@@ -1,4 +1,4 @@
+/* eslint-disable import/prefer-default-export */
 import { DateTime } from './luxon.js';
 
-const dateSection = document.getElementById('date');
-dateSection.innerText = DateTime.now().toLocaleString(DateTime.DATETIME_MED);
+export const date = DateTime.now().toLocaleString(DateTime.DATETIME_MED);
