--- modules/hideComponent.js
@@ -1,4 +1,0 @@
-const addNewSection = document.getElementById('add-new');
-const contactSection = document.getElementById('contact');
-addNewSection.style.display = 'none';
-contactSection.style.display = 'none';
